# user-management/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: TenantsService API
    version: 0.0.1
paths:
    /api/v1/um/change_password:
        post:
            tags:
                - TenantsService
            operationId: TenantsService_ChangePassword
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/tenants.v1.ChangePasswordRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/tenants.v1.ChangePasswordResponse'
    /api/v1/um/create_sub_user:
        post:
            tags:
                - TenantsService
            operationId: TenantsService_CreateSubUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/tenants.v1.CreateSubUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/tenants.v1.CreateSubUserResponse'
    /api/v1/um/delete_master_user/{identityId}:
        delete:
            tags:
                - TenantsService
            operationId: TenantsService_DeleteMasterUser
            parameters:
                - name: identityId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/tenants.v1.DeleteUserResponse'
    /api/v1/um/delete_sub_user/{identityId}:
        delete:
            tags:
                - TenantsService
            operationId: TenantsService_DeleteSubUser
            parameters:
                - name: identityId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/tenants.v1.DeleteUserResponse'
    /api/v1/um/forgot_password:
        post:
            tags:
                - TenantsService
            operationId: TenantsService_ForgotPassword
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/tenants.v1.ForgotPasswordRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/tenants.v1.ForgotPasswordResponse'
    /api/v1/um/list_all_users:
        get:
            tags:
                - TenantsService
            operationId: TenantsService_ListSubUsers
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/tenants.v1.ListSubUsersResponse'
    /api/v1/um/login:
        post:
            tags:
                - TenantsService
            operationId: TenantsService_LoginUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/tenants.v1.LoginUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/tenants.v1.LoginUserResponse'
    /api/v1/um/register_master_user:
        post:
            tags:
                - TenantsService
            operationId: TenantsService_RegisterMasterUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/tenants.v1.RegisterMasterUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/tenants.v1.RegisterMasterUserResponse'
    /api/v1/um/reset_password:
        post:
            tags:
                - TenantsService
            operationId: TenantsService_ResetPassword
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/tenants.v1.ResetPasswordRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/tenants.v1.ResetPasswordResponse'
    /api/v1/um/token:
        get:
            tags:
                - TenantsService
            operationId: TenantsService_GetJWTToken
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/tenants.v1.GetJWTTokenResponse'
    /api/v1/um/update_profile:
        post:
            tags:
                - TenantsService
            operationId: TenantsService_UpdateUserProfile
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/tenants.v1.UpdateUserProfileRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/tenants.v1.UpdateUserProfileResponse'
components:
    schemas:
        tenants.v1.ChangePasswordRequest:
            type: object
            properties:
                newPassword:
                    type: string
        tenants.v1.ChangePasswordResponse:
            type: object
            properties:
                status:
                    type: string
                message:
                    type: string
        tenants.v1.CreateSubUserRequest:
            type: object
            properties:
                email:
                    type: string
                password:
                    type: string
                firstName:
                    type: string
                lastName:
                    type: string
        tenants.v1.CreateSubUserResponse:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/tenants.v1.UserData'
        tenants.v1.DeleteUserResponse:
            type: object
            properties:
                status:
                    type: string
                message:
                    type: string
        tenants.v1.ForgotPasswordRequest:
            type: object
            properties:
                email:
                    type: string
        tenants.v1.ForgotPasswordResponse:
            type: object
            properties:
                status:
                    type: string
                message:
                    type: string
                flowId:
                    type: string
        tenants.v1.GetJWTTokenResponse:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/tenants.v1.TokenData'
        tenants.v1.ListSubUsersResponse:
            type: object
            properties:
                users:
                    type: array
                    items:
                        $ref: '#/components/schemas/tenants.v1.UserSummary'
        tenants.v1.LoginUserRequest:
            type: object
            properties:
                email:
                    type: string
                password:
                    type: string
        tenants.v1.LoginUserResponse:
            type: object
            properties:
                sessionToken:
                    type: string
                user:
                    $ref: '#/components/schemas/tenants.v1.UserInfo'
        tenants.v1.RegisterMasterUserRequest:
            type: object
            properties:
                email:
                    type: string
                password:
                    type: string
                firstName:
                    type: string
                lastName:
                    type: string
                organizationName:
                    type: string
        tenants.v1.RegisterMasterUserResponse:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/tenants.v1.TenantData'
        tenants.v1.ResetPasswordRequest:
            type: object
            properties:
                flowId:
                    type: string
                code:
                    type: string
                newPassword:
                    type: string
                email:
                    type: string
        tenants.v1.ResetPasswordResponse:
            type: object
            properties:
                status:
                    type: string
                message:
                    type: string
        tenants.v1.TenantData:
            type: object
            properties:
                identityId:
                    type: string
                organizationName:
                    type: string
                status:
                    type: string
        tenants.v1.TokenData:
            type: object
            properties:
                jwtToken:
                    type: string
        tenants.v1.UpdateUserProfileRequest:
            type: object
            properties:
                identityId:
                    type: string
                firstName:
                    type: string
                lastName:
                    type: string
                email:
                    type: string
                role:
                    type: string
        tenants.v1.UpdateUserProfileResponse:
            type: object
            properties:
                status:
                    type: string
                message:
                    type: string
                user:
                    $ref: '#/components/schemas/tenants.v1.UserInfo'
        tenants.v1.UserData:
            type: object
            properties:
                identityId:
                    type: string
                organizationName:
                    type: string
        tenants.v1.UserInfo:
            type: object
            properties:
                email:
                    type: string
                firstName:
                    type: string
                lastName:
                    type: string
                organizationName:
                    type: string
                role:
                    type: string
        tenants.v1.UserSummary:
            type: object
            properties:
                identityId:
                    type: string
                email:
                    type: string
                firstName:
                    type: string
                lastName:
                    type: string
                organizationName:
                    type: string
                role:
                    type: string
tags:
    - name: TenantsService